Added Clubs Basics

+++ uhill/database/migrations/2022_07_04_042210_create_courses_table.php
@@ -19,10 +19,10 @@ return new class extends Migration
             $table->foreignId('teacher_id')->constrained()->cascadeOnDelete();
             $table->string('subject');
             $table->integer('grade')->default('13');
-            $table->integer('personality')->default('5');
-            $table->integer('fairness')->default('5');
-            $table->integer('easiness')->default('5');
-            $table->integer('overall')->default('1');
+            $table->double('personality')->default('5');
+            $table->double('fairness')->default('5');
+            $table->double('easiness')->default('5');
+            $table->double('overall')->default('1');
             $table->integer('review_count')->default('0');
             $table->timestamp('date_added');
             $table->longText('description');
